requirements.txt
Django
psycopg2-binary
gunicorn
redis
python-dotenv
django-cors-headers
dj-database-url